+++ java/connector-node/risingwave-connector-test/src/test/resources/log4j2.properties
