+++ java/connector-node/risingwave-connector-test/src/test/resources/stored_queries.properties
